+++ backend/src/user/user.controller.ts
@@ -11,16 +11,21 @@ import {
 } from '@nestjs/common';
 import { Request } from 'express';
 
-import { UserGuard } from './user.guard';
+import { AcceptFriendGuard } from './guard/accept-friend.guard';
+import { BlockedUserGuard } from './guard/blocked-user.guard';
+import { CreateFriendRequestGuard } from './guard/create-friend-request.guard';
+import { DeleteFriendGuard } from './guard/delete-friend.guard';
+import { DeleteBlockGuard } from './guard/delete-block.guard';
+import { UserExistGuard } from './guard/user-exist.guard';
 import { UserId, VerifiedRequest } from '../util/type';
 import { UserService } from './user.service';
 
 @Controller('user')
-@UseGuards(UserGuard)
 export class UserController {
   constructor(private readonly userService: UserService) {}
 
   @Get(':userId/info')
+  @UseGuards(UserExistGuard)
   findProfile(@Req() req: Request, @Param(':userId') userId: UserId) {}
 
   /*****************************************************************************
@@ -30,9 +35,11 @@ export class UserController {
    ****************************************************************************/
 
   @Post(':userId/game')
+  @UseGuards(UserExistGuard, BlockedUserGuard)
   createGame(@Req() req: Request, @Param(':userId') userId: UserId) {}
 
   @Get(':userId/game/:gameId')
+  @UseGuards(UserExistGuard, BlockedUserGuard)
   findGame(@Req() req: Request, @Param(':userId') userId: UserId) {}
 
   /*****************************************************************************
@@ -42,9 +49,11 @@ export class UserController {
    ****************************************************************************/
 
   @Put(':userId/block')
+  @UseGuards(UserExistGuard, BlockedUserGuard)
   createBlock(@Req() req: Request, @Param(':userId') userId: UserId) {}
 
   @Delete(':userId/block')
+  @UseGuards(UserExistGuard, BlockedUserGuard, DeleteBlockGuard)
   deleteBlock(@Req() req: Request, @Param(':userId') userId: UserId) {}
 
   /*****************************************************************************
@@ -63,11 +72,14 @@ export class UserController {
   }
 
   @Put(':userId/friend')
+  @UseGuards(UserExistGuard, BlockedUserGuard, CreateFriendRequestGuard)
   createFriendRequest(@Req() req: Request, @Param(':userId') userId: UserId) {}
 
   @Delete(':userId/friend')
+  @UseGuards(UserExistGuard, BlockedUserGuard, DeleteFriendGuard)
   deleteFriendship(@Req() req: Request, @Param(':userId') userId: UserId) {}
 
   @Patch(':userId/friend')
+  @UseGuards(UserExistGuard, BlockedUserGuard, AcceptFriendGuard)
   updateFriendship(@Req() req: Request, @Param(':userId') userId: UserId) {}
 }
